requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
cloudinary==1.37.0
openai==1.3.0